requirements.txt
anthropic
chromadb
python-dotenv